Add files via upload
This Project is focused on conducting exploratory data analysis on a dataset related to Netflix content. The primary objective of this project is
to uncover valuable insights, detect trends, and draw conclusions from the data. To accomplish this, the dataset is analyzed using various
techniques, including data cleaning, visualizations, and statistical analysis. The ultimate goal is to gain a comprehensive understanding of the
data, which can aid in making informed decisions and formulating effective strategies.

Analyzed Netflix's extensive dataset, using Python and data analysis techniques to uncover valuable insights
that helped strategize content delivery optimization.